e2e test now tests by passing cmd args to top-level run().
