+++ helpers_test.go
@@ -1,4 +1,4 @@
-package command
+package main
 
 import (
 	"context"
